# Conflicts:
#	API/Data/DataContext.cs
#	API/Data/Migrations/DataContextModelSnapshot.cs
#	API/Data/Seed.cs
#	API/Data/UnitOfWork.cs
#	API/Extensions/ApplicationServiceExtensions.cs
#	API/Helpers/AutoMapperProfiles.cs
#	API/I18N/en.json
#	API/Services/DirectoryService.cs
#	API/Services/Tasks/Scanner/Parser/Parser.cs
#	UI/Web/src/app/book-reader/_components/book-reader/book-reader.component.scss
#	UI/Web/src/app/book-reader/_components/book-reader/book-reader.component.ts
#	UI/Web/src/app/book-reader/_components/reader-settings/reader-settings.component.html
#	UI/Web/src/app/book-reader/_components/reader-settings/reader-settings.component.ts
#	UI/Web/src/app/book-reader/_services/book.service.ts
#	UI/Web/src/app/user-settings/theme-manager/theme-manager.component.html
#	UI/Web/src/app/user-settings/user-preferences/user-preferences.component.html
#	UI/Web/src/app/user-settings/user-preferences/user-preferences.component.ts
#	UI/Web/src/assets/langs/en.json
#	openapi.json

+++ API/Controllers/ThemeController.cs
@@ -103,7 +103,7 @@ public class ThemeController : BaseApiController
     [HttpDelete]
     public async Task<ActionResult<IEnumerable<DownloadableSiteThemeDto>>> DeleteTheme(int themeId)
     {
-
+        if (User.IsInRole(PolicyConstants.ReadOnlyRole)) return BadRequest(await _localizationService.Translate(User.GetUserId(), "permission-denied"));
         await _themeService.DeleteTheme(themeId);
 
         return Ok();
@@ -128,6 +128,8 @@ public class ThemeController : BaseApiController
     [HttpPost("upload-theme")]
     public async Task<ActionResult<SiteThemeDto>> DownloadTheme(IFormFile formFile)
     {
+        if (User.IsInRole(PolicyConstants.ReadOnlyRole)) return BadRequest(await _localizationService.Translate(User.GetUserId(), "permission-denied"));
+
         if (!formFile.FileName.EndsWith(".css")) return BadRequest("Invalid file");
         if (formFile.FileName.Contains("..")) return BadRequest("Invalid file");
         var tempFile = await UploadToTemp(formFile);
