+++ UI/Web/src/app/book-reader/_components/personal-table-of-contents/personal-table-of-contents.component.ts
@@ -1,19 +1,22 @@
 import {
   ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component, DestroyRef, EventEmitter,
-  Inject,
+  Component,
+  DestroyRef,
+  EventEmitter,
   inject,
   Input,
+  model,
   OnInit,
   Output
 } from '@angular/core';
-import {CommonModule, DOCUMENT} from '@angular/common';
 import {ReaderService} from "../../../_services/reader.service";
 import {PersonalToC} from "../../../_models/readers/personal-toc";
 import {takeUntilDestroyed} from "@angular/core/rxjs-interop";
-import {NgbTooltip} from "@ng-bootstrap/ng-bootstrap";
-import {TranslocoDirective} from "@ngneat/transloco";
+import {translate, TranslocoDirective} from "@jsverse/transloco";
+import {TextBookmarkItemComponent} from "../text-bookmark-item/text-bookmark-item.component";
+import {ConfirmService} from "../../../shared/confirm.service";
+import {FilterPipe} from "../../../_pipes/filter.pipe";
+import {FormControl, FormGroup, FormsModule, ReactiveFormsModule} from "@angular/forms";
 
 export interface PersonalToCEvent {
   pageNum: number;
@@ -22,31 +25,29 @@ export interface PersonalToCEvent {
 
 @Component({
   selector: 'app-personal-table-of-contents',
-  standalone: true,
-  imports: [CommonModule, NgbTooltip, TranslocoDirective],
+  imports: [TranslocoDirective, TextBookmarkItemComponent, FilterPipe, FormsModule, ReactiveFormsModule],
   templateUrl: './personal-table-of-contents.component.html',
   styleUrls: ['./personal-table-of-contents.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PersonalTableOfContentsComponent implements OnInit {
 
+  private readonly readerService = inject(ReaderService);
+  private readonly destroyRef = inject(DestroyRef);
+  private readonly confirmService = inject(ConfirmService);
+
+  protected readonly ShowFilterAfterItems = 10;
+
   @Input({required: true}) chapterId!: number;
   @Input({required: true}) pageNum: number = 0;
   @Input({required: true}) tocRefresh!: EventEmitter<void>;
   @Output() loadChapter: EventEmitter<PersonalToCEvent> = new EventEmitter();
 
-  private readonly readerService = inject(ReaderService);
-  private readonly cdRef = inject(ChangeDetectorRef);
-  private readonly destroyRef = inject(DestroyRef);
 
-
-  bookmarks: {[key: number]: Array<PersonalToC>} = [];
-
-  get Pages() {
-    return Object.keys(this.bookmarks).map(p => parseInt(p, 10));
-  }
-
-  constructor(@Inject(DOCUMENT) private document: Document) {}
+  ptocBookmarks = model<PersonalToC[]>([]);
+  formGroup = new FormGroup({
+    filter: new FormControl('', [])
+  });
 
   ngOnInit() {
     this.tocRefresh.pipe(takeUntilDestroyed(this.destroyRef)).subscribe(() => {
@@ -58,13 +59,7 @@ export class PersonalTableOfContentsComponent implements OnInit {
 
   load() {
     this.readerService.getPersonalToC(this.chapterId).subscribe(res => {
-      res.forEach(t => {
-        if (!this.bookmarks.hasOwnProperty(t.pageNumber)) {
-          this.bookmarks[t.pageNumber] = [];
-        }
-        this.bookmarks[t.pageNumber].push(t);
-      })
-      this.cdRef.markForCheck();
+      this.ptocBookmarks.set(res);
     });
   }
 
@@ -72,15 +67,18 @@ export class PersonalTableOfContentsComponent implements OnInit {
     this.loadChapter.emit({pageNum, scrollPart});
   }
 
-  removeBookmark(bookmark: PersonalToC) {
-    this.readerService.removePersonalToc(bookmark.chapterId, bookmark.pageNumber, bookmark.title).subscribe(() => {
-      this.bookmarks[bookmark.pageNumber] = this.bookmarks[bookmark.pageNumber].filter(t => t.title != bookmark.title);
+  async removeBookmark(bookmark: PersonalToC) {
 
-      if (this.bookmarks[bookmark.pageNumber].length === 0) {
-        delete this.bookmarks[bookmark.pageNumber];
-      }
-      this.cdRef.markForCheck();
+    if (!await this.confirmService.confirm(translate('toasts.confirm-delete-bookmark'))) return;
+
+    this.readerService.removePersonalToc(bookmark.chapterId, bookmark.pageNumber, bookmark.title).subscribe(() => {
+      this.ptocBookmarks.set(this.ptocBookmarks().filter(t => t.title !== bookmark.title));
     });
+  }
+
+  filterList = (listItem: PersonalToC) => {
+    const query = (this.formGroup.get('filter')?.value || '').toLowerCase();
+    return listItem.title.toLowerCase().indexOf(query) >= 0 || listItem.pageNumber.toString().indexOf(query) >= 0 || (listItem.chapterTitle ?? '').toLowerCase().indexOf(query) >= 0;
   }
 
 }
